+++ packages/client/modules/userDashboard/components/GcalModal/DateTimePickers.tsx
@@ -42,38 +42,17 @@ const timePickerStyles = {
 
 type Props = {
   startValue: Dayjs
-  setStart: (newValue: Dayjs) => void
   endValue: Dayjs
-  setEnd: (newValue: Dayjs) => void
+  handleChangeStart: (date: Dayjs | null, time: Dayjs | null) => void
+  handleChangeEnd: (date: Dayjs | null, time: Dayjs | null) => void
 }
 
 const DateTimePickers = (props: Props) => {
-  const {startValue, setStart, endValue, setEnd} = props
+  const {startValue, endValue, handleChangeStart, handleChangeEnd} = props
   const timeZone = Intl.DateTimeFormat().resolvedOptions().timeZone
   const date = new Date()
   const dateTimeString = date.toLocaleString('en-US', {timeZone: timeZone, timeZoneName: 'short'})
   const timeZoneShort = dateTimeString.split(' ').pop()
-
-  const handleChangeStart = (date: Dayjs | null, time: Dayjs | null) => {
-    if (date && time) {
-      const newValue = date.hour(time.hour()).minute(time.minute())
-      setStart(newValue)
-      setEnd(newValue.add(1, 'hour'))
-    }
-  }
-
-  const handleChangeEnd = (date: Dayjs | null, time: Dayjs | null) => {
-    if (date && time) {
-      const newValue = date.hour(time.hour()).minute(time.minute())
-      if (newValue.isAfter(startValue)) {
-        setEnd(newValue)
-      } else {
-        const newStartValue = newValue.subtract(1, 'hour')
-        setStart(newStartValue)
-        setEnd(newValue)
-      }
-    }
-  }
 
   const handleMouseDown = (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
     // prevent gcal modal from closing when clicking datetime pickers
